requirements.txt
Appium-Python-Client
selenium
pytest
invoke